refactor(*): Change files location
