--- src/App.style.css
@@ -1,3 +1,0 @@
-.main-section {
-  padding: 0;
-  margin: 0;
