--- src/components/Card/Card.style.css
@@ -1,90 +1,0 @@
-.card {
-  width: 340px;
-  height: 350px;
-  background-color: white;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-  padding: 0;
-  border-radius: 15px;
-  overflow: hidden;
-}
-
-.card:hover {
-  border: 3px solid var(--color-primary);
-}
-
-.card-img-section {
-  width: 340px;
-  height: 200px;
-  border-top-left-radius: 15px;
-  border-top-right-radius: 15px;
-  overflow: hidden;
-}
-
-.no-card-list {
-  height: 500px;
-  padding: 5rem;
-}
-
-.sample-img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.sample-img-bigger {
-  transform: scale(1.5);
-  transition: transform 0.3s;
-}
-
-.card-text-section {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: left;
-  padding: 20px 15px;
-  gap: 10px;
-}
-.time-stamp {
-  color: #666;
-  font-family: Pretendard;
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  margin: 0;
-  padding: 0;
-}
-
-.introduce-text {
-  margin: 0;
-  padding: 0;
-  width: 300px;
-  height: 50px;
-  color: #000;
-  text-overflow: ellipsis;
-  font-family: Pretendard;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 24px; /* 150% */
-  white-space: normal; /*글자를 한줄로 모아준다*/
-  overflow: hidden;
-  word-break: break-word;
-
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-}
-
-.created-date {
-  margin: 0;
-  padding: 0;
-  color: #333;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  font-family: Pretendard;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
