--- src/components/Footer/Footer.style.css
@@ -1,46 +1,0 @@
-footer {
-  background: var(--the-julge-black, var(--color-black));
-  padding: 3rem 3rem 10rem;
-  position: relative;
-}
-.content {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: 1.6rem;
-  font-weight: 400;
-}
-.footer.logo {
-  color: #676767;
-  text-align: center;
-  position: absolute;
-  bottom: 3rem;
-}
-.footer.link {
-  display: flex;
-  gap: 3rem;
-  justify-content: center;
-  align-items: center;
-}
-.footer.link a {
-  color: #cfcfcf;
-}
-.footer.sns {
-  display: flex;
-  gap: 1.2rem;
-  justify-content: space-between;
-  align-items: center;
-}
-
-/* 768px 이상 1199px 이하 타블렛에서는 
- * 
- * footer의 padding을 조절한다.
-*/
-@media (min-width: 768px) {
-  footer {
-    padding: 3rem 10rem 10rem;
-  }
-  .footer.logo {
-    position: static;
-  }
-}
