--- src/components/Nav/Nav.style.css
@@ -1,64 +1,0 @@
-/* nav css 작성 */
-nav {
-  width: 100%;
-  background-color: var(--color-gray1);
-  padding: 2rem 3.2rem;
-  position: sticky;
-  top: 0;
-  left: 0;
-  z-index: 10;
-  display: flex;
-  flex-wrap: nowrap;
-  justify-content: center;
-  align-items: center;
-}
-.gnb {
-  width: 100%;
-  padding: 0;
-  display: flex;
-  flex-wrap: nowrap;
-  justify-content: space-between;
-  align-items: center;
-}
-.link-button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 1.6rem 2rem;
-  border: none;
-  border-radius: 8px;
-  background: var(--gra-purpleblue-to-skyblue);
-}
-.link-button .a-label {
-  color: var(--grey-light);
-  font-size: 1.8rem;
-  font-weight: 600;
-  padding: 0 1rem;
-}
-.user-info {
-  display: flex;
-  gap: 1rem;
-}
-.profile-img {
-  width: 1.5rem;
-  height: 1.5rem;
-  border-radius: 70%;
-  overflow: hidden;
-}
-
-/* 768px 이상 1199px 이하 타블렛에서는 
- * nav 좌우 padding이 최소 3.2rem이며 gnb div의 크기가 최대 80rem이었다가 nav 태그의 크기가 86.4rem보다 작아지면 gnb의 로고와 버튼 사이 간격을 줄인다.
-*/
-@media (min-width: 768px) {
-  .gnb {
-    max-width: 80rem;
-  }
-}
-/* 1200px 이상 컴퓨터에서는
- * 좌우 padding이 20rem이며 브라우저 창이 작아지면 로고와 버튼 간의 간격이 줄어든다.
- */
-@media (min-width: 1200px) {
-  nav {
-    padding: 2rem 10rem;
-  }
-}
