--- src/components/SearchBar/SearchBar.style.css
@@ -1,24 +1,0 @@
-.search-form {
-  /* width: 100%; */
-  position: relative;
-}
-
-.search-bar {
-  padding: 1rem 4rem;
-  width: 100%;
-  height: 3.5rem;
-  border-radius: 10px;
-  border: none;
-  background-color: var(--color-gray2);
-}
-
-input:focus {
-  outline: none !important;
-  border: 2px solid var(--color-gray4);
-}
-
-.search-icon {
-  position: absolute;
-  top: 1rem;
-  left: 1rem;
-}
