+++ src/index.js
@@ -1,4 +1,4 @@
 import ReactDOM from "react-dom";
-import Main from "./Main";
+import Main from "Main";
 
 ReactDOM.render(<Main />, document.getElementById("root"));
