+++ src/pages/SharedPage/components/ShareHeader/ShareHeader.js
